perk.h: eliminate #include's of referential headers.
The primary purpose of this commit is to eliminate the pulling
in of definitions not strictly needed by perk.h. In particular,
perk.h does not require any of the PE constants to be defined,
and only refers to the raw PE structs by pointer.

The secondary purpose of this commit is to allow for inclusion
of perk.h by software that already comes with its own set of
PE const and struct definitions.